requirements.txt
pdfminer.six==20220319
Pillow>=9.1
Wand>=0.6.7